Explicit cast to avoid warning with g++ 4.3.3
git-svn-id: https://devel.codership.com/svn/galera/trunk@1042 9148e567-b586-4176-b9b8-4e36cb4ec61b

+++ galerautils/src/gu_time.h
@@ -25,7 +25,7 @@ static inline void
 gu_timeval_add (struct timeval* time, double s)
 {
     long sec  = (long)s;
-    long usec = (long)((s - sec) * 1000000);
+    long usec = (long)(((long)s - sec) * 1000000);
     long carry;
     time->tv_usec += usec;
     carry = (time->tv_usec >= 1000000);
